functions/raw-fetch-fred/requirements.txt
functions-framework==3.*
google-cloud-secret-manager
duckdb==1.3.2
pandas
pyarrow
requests
google-cloud-storage
google-cloud-bigquery
pyyaml
tqdm